requirements.txt
python-telegram-bot[all]>=20.0
psutil>=5.9.0
GPUtil>=1.4.0
speedtest-cli>=2.1.3
humanize>=4.6.0
yt-dlp>=2023.12.30
aiohttp>=3.9.1
